requirements.txt
keyboard
soundfile
sounddevice
novelai_api
numpy
webrtcvad
openai-whisper
aiohttp
python-dotenv